requirements.txt
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cmake==3.22.4
coverage==5.5
deprecation==2.1.0
dlib==19.24.0
eyed3==0.9.6
face-recognition==1.3.0
face-recognition-models==0.3.0
filetype==1.0.13
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQt5-stubs==5.15.6.0
toml==0.10.2
python-dotenv~=0.20.0